// src/generated/models/EntitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Entitiescomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'Deleted Unpublished'
} as const;
export type Entitiescomponentstate = keyof typeof Entitiescomponentstate;

export interface Entities {
  addresstablename?: string;
  basetablename?: string;
  collectionname?: string;
  componentstate: Entitiescomponentstate;
  entityid: string;
  entitysetname?: string;
  extensiontablename?: string;
  externalcollectionname?: string;
  externalname?: string;
  isactivity?: boolean;
  isactivityname?: string;
  logicalcollectionname?: string;
  logicalname?: string;
  name?: string;
  objecttypecode?: number;
  originallocalizedcollectionname?: string;
  originallocalizedname?: string;
  overwritetime: string;
  parentcontrollingattributename?: string;
  physicalname?: string;
  reportviewname?: string;
  solutionid: string;
  versionnumber?: number;
}