// src/generated/models/SystemformsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Systemformscomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'DeletedUnpublished'
} as const;
export type Systemformscomponentstate = keyof typeof Systemformscomponentstate;
export const Systemformsformactivationstate = {
  0: 'Inactive',
  1: 'Active'
} as const;
export type Systemformsformactivationstate = keyof typeof Systemformsformactivationstate;
export const Systemformsformpresentation = {
  0: 'ClassicForm',
  1: 'AirForm',
  2: 'ConvertedICForm'
} as const;
export type Systemformsformpresentation = keyof typeof Systemformsformpresentation;
export const Systemformstype = {
  0: 'Dashboard',
  1: 'AppointmentBook',
  2: 'Main',
  3: 'MiniCampaignBO',
  4: 'Preview',
  5: 'Mobile_Express',
  6: 'QuickViewForm',
  7: 'QuickCreate',
  8: 'Dialog',
  9: 'TaskFlowForm',
  10: 'InteractionCentricDashboard',
  11: 'Card',
  12: 'Main_Interactiveexperience',
  13: 'ContextualDashboard',
  100: 'Other',
  101: 'MainBackup',
  102: 'AppointmentBookBackup',
  103: 'PowerBIDashboard'
} as const;
export type Systemformstype = keyof typeof Systemformstype;

export interface SystemformsBase {
  "AncestorFormId@odata.bind"?: string;
  canbedeleted: string;
  description?: string;
  formactivationstate: Systemformsformactivationstate;
  formid: string;
  formjson: string;
  formpresentation: Systemformsformpresentation;
  formxml: string;
  introducedversion?: string;
  isairmerged?: boolean;
  iscustomizable: string;
  isdefault: boolean;
  isdesktopenabled: boolean;
  istabletenabled: boolean;
  name: string;
  objecttypecode?: string;
  type?: Systemformstype;
  uniquename?: string;
  version?: number;
}

export interface Systemforms extends SystemformsBase {
  ancestorformidname?: string;
  componentstate: Systemformscomponentstate;
  formactivationstatename?: string;
  formidunique: string;
  formpresentationname?: string;
  formxmlmanaged?: string;
  isdefaultname?: string;
  ismanaged: boolean;
  ismanagedname?: string;
  objecttypecodename?: string;
  organizationidname: string;
  overwritetime: string;
  publishedon?: string;
  solutionid: string;
  supportingsolutionid?: string;
  typename?: string;
  versionnumber?: number;
  ancestorformid?: object;
  _ancestorformid_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
}